+++ weave/conftest.py
@@ -1,4 +1,3 @@
-import base64
 import logging
 import os
 import pathlib
@@ -20,7 +19,6 @@ from weave.legacy import context_state, io_service, serialize
 from weave.legacy.language_features.tagging.tag_store import isolated_tagging_context
 from weave.trace_server import (
     clickhouse_trace_server_batched,
-    external_to_internal_trace_server_adapter,
     remote_http_trace_server,
     sqlite_trace_server,
 )
@@ -301,112 +299,8 @@ def strict_op_saving():
 #     )
 
 
-class TwoWayMapping:
-    def __init__(self):
-        self._ext_to_int_map = {}
-        self._int_to_ext_map = {}
-
-        # Useful for testing to ensure caching is working
-        self.stats = {
-            "ext_to_int": {
-                "hits": 0,
-                "misses": 0,
-            },
-            "int_to_ext": {
-                "hits": 0,
-                "misses": 0,
-            },
-        }
-
-    def ext_to_int(self, key, default=None):
-        if key not in self._ext_to_int_map:
-            if default is None:
-                raise ValueError(f"Key {key} not found")
-            if default in self._int_to_ext_map:
-                raise ValueError(f"Default {default} already in use")
-            self._ext_to_int_map[key] = default
-            self._int_to_ext_map[default] = key
-            self.stats["ext_to_int"]["misses"] += 1
-        else:
-            self.stats["ext_to_int"]["hits"] += 1
-        return self._ext_to_int_map[key]
-
-    def int_to_ext(self, key, default):
-        if key not in self._int_to_ext_map:
-            if default is None:
-                raise ValueError(f"Key {key} not found")
-            if default in self._ext_to_int_map:
-                raise ValueError(f"Default {default} already in use")
-            self._int_to_ext_map[key] = default
-            self._ext_to_int_map[default] = key
-            self.stats["int_to_ext"]["misses"] += 1
-        else:
-            self.stats["int_to_ext"]["hits"] += 1
-        return self._int_to_ext_map[key]
-
-
-def b64(s: str) -> str:
-    # Base64 encode the string
-    return base64.b64encode(s.encode("ascii")).decode("ascii")
-
-
-class DummyIdConverter(external_to_internal_trace_server_adapter.IdConverter):
-    def __init__(self):
-        self._project_map = TwoWayMapping()
-        self._run_map = TwoWayMapping()
-        self._user_map = TwoWayMapping()
-
-    def ext_to_int_project_id(self, project_id: str) -> str:
-        return self._project_map.ext_to_int(project_id, b64(project_id))
-
-    def int_to_ext_project_id(self, project_id: str) -> typing.Optional[str]:
-        return self._project_map.int_to_ext(project_id, b64(project_id))
-
-    def ext_to_int_run_id(self, run_id: str) -> str:
-        return self._run_map.ext_to_int(run_id, b64(run_id) + ":" + run_id)
-
-    def int_to_ext_run_id(self, run_id: str) -> str:
-        exp = run_id.split(":")[1]
-        return self._run_map.int_to_ext(run_id, exp)
-
-    def ext_to_int_user_id(self, user_id: str) -> str:
-        return self._user_map.ext_to_int(user_id, b64(user_id))
-
-    def int_to_ext_user_id(self, user_id: str) -> str:
-        return self._user_map.int_to_ext(user_id, b64(user_id))
-
-
-class TestOnlyUserInjectingExternalTraceServer(
-    external_to_internal_trace_server_adapter.ExternalTraceServer
-):
-    def __init__(
-        self,
-        internal_trace_server: tsi.TraceServerInterface,
-        id_converter: external_to_internal_trace_server_adapter.IdConverter,
-        user_id: str,
-    ):
-        super().__init__(internal_trace_server, id_converter)
-        self._user_id = user_id
-
-    def call_start(self, req: tsi.CallStartReq) -> tsi.CallStartRes:
-        req.start.wb_user_id = self._user_id
-        return super().call_start(req)
-
-    def calls_delete(self, req: tsi.CallsDeleteReq) -> tsi.CallsDeleteRes:
-        req.wb_user_id = self._user_id
-        return super().calls_delete(req)
-
-    def call_update(self, req: tsi.CallUpdateReq) -> tsi.CallUpdateRes:
-        req.wb_user_id = self._user_id
-        return super().call_update(req)
-
-    def feedback_create(self, req: tsi.FeedbackCreateReq) -> tsi.FeedbackCreateRes:
-        req.wb_user_id = self._user_id
-        return super().feedback_create(req)
-
-
 @pytest.fixture()
-def init_client(request) -> Generator[InitializedClient, None, None]:
+def client(request) -> Generator[weave_client.WeaveClient, None, None]:
     inited_client = None
     weave_server_flag = request.config.getoption("--weave-server")
     server: tsi.TraceServerInterface
@@ -444,14 +338,9 @@ def init_client(request) -> Generator[InitializedClient, None, None]:
         inited_client = weave_init.InitializedClient(client)
         autopatch.autopatch()
     try:
-        yield inited_client
+        yield inited_client.client
     finally:
         inited_client.reset()
-
-
-@pytest.fixture()
-def client(init_client) -> Generator[weave_client.WeaveClient, None, None]:
-    yield init_client.client
 
 
 @pytest.fixture
